Moved anticheat.py
